Refactored code.

+++ src/lexer.c
@@ -60,7 +60,7 @@ static bool isSigned(struct list *list, const char *str, const int i)
     }
 
     return
-        ((int)getType(str[i-1]) == Operator && str[i-1] != ')' && str[i-1] != '!') &&
+        (getType(str[i-1]) == Operator && str[i-1] != ')' && str[i-1] != '!') &&
         (list->last == NULL || !isNumber(list->last->value));
 }
